requirements.txt
python-telegram-bot==20.0
httpx==0.27.0
python-dotenv==1.0.1
spacy==3.7.4
pytesseract==0.3.10
Pillow==10.2.0
textract==1.6.5
langdetect==1.0.9
docx2txt==0.8
pdf2image==1.17.0